requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
pandas
openpyxl
qdrant-client
openai
python-dotenv
